Changed axios config location
